AI Fusion/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
Django==4.2.7
google==3.0.0
google-trends==1.1
gTTS==2.4.0
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
lxml==4.9.3
moviepy==1.0.3
numpy==1.26.2
opencv-python==4.8.1.78
pandas==2.1.3
Pillow==10.0.1
proglog==0.1.10
pyee==11.0.1
python-dateutil==2.8.2
python-ffmpeg==2.0.4
pytrends==4.9.2
pytube==15.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
SpeechRecognition==3.10.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
